refactor(s2dm): Split exporter code into multiple manageable files
- Added `metadata_tracker.py` for initializing and managing VSS metadata.
- Created `reference_generator.py` to handle VSS reference file generation alongside GraphQL schema output.
- Developed `schema_generator.py` for orchestrating the complete S2DM GraphQL schema generation process.
- Introduced `type_builders.py` for constructing GraphQL types from VSS elements, including unit enums, instance types, and struct types.
- Updated tests to reflect changes in module imports and ensure functionality of new features.
